Move load_from_ultralytics into _checkpoint.py
